Rewrite template and style
Part of the goal of the design of Google Play Music Web Manager is
to embed neatly into the official Google Play Music web interface.
The previous design assumed that we had control over the entire
div#music-content, but replacing its contents with an iframe appears
to break the web interface.  Better to inject the iframe inside
div.upload-music-view, which already has the necessary margins,
and fix the iframe's height after the fact.