Use new TrackList field names in template
"Last updated" was also changed to "last modified", because it's
marginally more descriptive - what the timestamp actually communicates
is the most recent time one of the tracks in the listing was modified.
Deleting a track, for instance, can thus be used to roll the timestamp
back, which is a bit counterintuitive.