Bring back the upload form
Turns out there's no pure-HTML way to do pure-binary POSTs or PUTs
in a browser.  I had hesitated to do this before, because I didn't
want to add a raw upload method to musicmanager.Client, but I also
couldn't figure out a way to get an urlfetch-capable *http.Client
without adding yet another appengine/standalone-specific function.
(oauth2.NewClient with a nil TokenSource didn't work, despite the
docs.)  However, turns out it's as easy as creating an *http.Client
with Transport set to an *urlfetch.Transport, which required only
converting fixTransport into getTransport.

Hopefully Google adds a more flexible method of doing out-of-request
processing to App Engine.  I'd like to make it possible to upload
multiple tracks at a time and receive a response the moment they've
all been cleared for upload, but goroutines are terminated when the
request handler quits, and appengine/delay only supports a kilobyte's
worth of arguments.  I'd have to dump the tracks in a Blobstore and
ask a delayed task to upload them from there, which means I'd also
have to implement a Blobstore shim for the standalone client...